requirements.txt
Flask==2.2.5
flask-socketio==5.3.2
python-socketio==5.9.0
eventlet==0.33.3
face_recognition==1.3.0
opencv-python==4.7.0.72
numpy>=1.24,<2.0
Pillow==10.0.0
gunicorn==20.1.0
pycryptodome==3.19.0
python-dotenv==1.0.0